supa_back/backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
supabase==2.5.0
pydantic==2.7.4
python-dotenv==1.0.1
openai==1.33.0
google-generativeai==0.6.0
pydantic-settings==2.3.1
xhtml2pdf==0.2.17
beautifulsoup4==4.12.3
pypdf==5.4.0
PyMuPDF==1.26.4
